chore: reusable table config between history content and change log
