+++ app/components/AnalystDashboard/ProjectChangeLog.tsx
@@ -19,22 +19,12 @@ import { diff } from 'json-diff';
 import { generateRawDiff, processArrayDiff } from 'components/DiffTable';
 import getConfig from 'next/config';
 import cbcData from 'formSchema/uiSchema/history/cbcData';
-import applicationDiffSchema from 'formSchema/uiSchema/history/application';
-import applicationGisDataSchema from 'formSchema/uiSchema/history/applicationGisData';
-import rfiDiffSchema from 'formSchema/uiSchema/history/rfi';
-import projectInformationSchema from 'formSchema/uiSchema/history/projectInformation';
-import conditionalApprovalSchema from 'formSchema/uiSchema/history/conditionalApproval';
-import screeningSchema from 'formSchema/uiSchema/history/screening';
-import gis from 'formSchema/uiSchema/history/gis';
-import gisAssessmentHhSchema from 'formSchema/uiSchema/history/gisAssessmentHh';
-import applicationSowDataSchema from 'formSchema/uiSchema/history/applicationSowData';
-import applicationAnnounced from 'formSchema/uiSchema/history/applicationAnnounced';
-import fnhaContribution from 'formSchema/uiSchema/history/fnhaContribution';
 import communities from 'formSchema/uiSchema/history/communities';
 import styled from 'styled-components';
 import { Box, Link, TableCellProps } from '@mui/material';
 import { DateTime } from 'luxon';
 import { processHistoryItems, formatUserName } from 'utils/historyProcessing';
+import { getTableConfig } from 'utils/historyTableConfig';
 import ClearFilters from 'components/Table/ClearFilters';
 import { getLabelForType } from 'components/Analyst/History/HistoryFilter';
 import AdditionalFilters from './AdditionalFilters';
@@ -161,248 +151,6 @@ const communityArrayToHistoryString = (
   keys: string[]
 ) =>
   communitiesArray.map((obj) => keys.map((key) => obj?.[key] ?? '').join(' '));
-// Define inline schemas for simple cases
-const communityReportSchema = {
-  communityReport: {
-    properties: {
-      dueDate: {
-        title: 'Due date',
-        type: 'string',
-      },
-      dateReceived: {
-        title: 'Date received',
-        type: 'string',
-      },
-    },
-  },
-};
-
-const pendingChangeRequestSchema = {
-  pendingChangeRequest: {
-    properties: {
-      comment: {
-        title: 'Comments:',
-        type: 'string',
-      },
-    },
-  },
-};
-
-const applicationDependenciesSchema = {
-  applicationDependencies: {
-    properties: {
-      crtcProjectDependent: { title: 'CRTC Project Dependent' },
-      connectedCoastNetworkDependent: {
-        title: 'Connected Coast Network Dependent',
-      },
-    },
-  },
-};
-
-// Map table names to their schemas and excluded keys
-const getTableConfig = (tableName: string, assessmentType?: string) => {
-  const configs = {
-    rfi_data: {
-      schema: rfiDiffSchema,
-      excludedKeys: [
-        'id',
-        'createdAt',
-        'updatedAt',
-        'applicationId',
-        'name',
-        'size',
-        'type',
-        'rfiEmailCorrespondance',
-        'fileDate',
-        'uploadedAt',
-        'eligibilityAndImpactsCalculator',
-        'detailedBudget',
-        'financialForecast',
-        'lastMileIspOffering',
-        'popWholesalePricing',
-        'communityRuralDevelopmentBenefitsTemplate',
-        'wirelessAddendum',
-        'supportingConnectivityEvidence',
-        'geographicNames',
-        'equipmentDetails',
-        'copiesOfRegistration',
-        'preparedFinancialStatements',
-        'logicalNetworkDiagram',
-        'projectSchedule',
-        'communityRuralDevelopmentBenefits',
-        'otherSupportingMaterials',
-        'geographicCoverageMap',
-        'coverageAssessmentStatistics',
-        'currentNetworkInfastructure',
-        'upgradedNetworkInfrastructure',
-        'uuid',
-      ],
-      overrideParent: 'rfi',
-    },
-    application_announced: {
-      schema: applicationAnnounced,
-      excludedKeys: [
-        'id',
-        'updated_at',
-        'created_at',
-        'created_by',
-        'updated_by',
-        'archived_at',
-        'archived_by',
-      ],
-      overrideParent: 'applicationAnnounced',
-    },
-    form_data: {
-      schema: applicationDiffSchema,
-      excludedKeys: [
-        'id',
-        'createdAt',
-        'updatedAt',
-        'applicationId',
-        'acknowledgements',
-        'supportingDocuments',
-        'coverage',
-        'templateUploads',
-      ],
-    },
-    application_dependencies: {
-      schema: applicationDependenciesSchema,
-      excludedKeys: ['id', 'createdAt', 'updatedAt', 'applicationId'],
-      overrideParent: 'applicationDependencies',
-    },
-    assessment_data: {
-      schema: (() => {
-        if (assessmentType === 'screening') return screeningSchema;
-        if (assessmentType === 'gis') return gis;
-        return null;
-      })(),
-      excludedKeys: [
-        'id',
-        'name',
-        'size',
-        'type',
-        'uuid',
-        'uploadedAt',
-        'otherFiles',
-        'assessmentTemplate',
-      ],
-      overrideParent: assessmentType,
-    },
-    conditional_approval_data: {
-      schema: conditionalApprovalSchema,
-      excludedKeys: [
-        'id',
-        'createdAt',
-        'updatedAt',
-        'uploadedAt',
-        'size',
-        'name',
-        'uuid',
-        'type',
-        'letterOfApprovalUpload',
-      ],
-      overrideParent: 'conditionalApproval',
-    },
-    application_gis_data: {
-      schema: applicationGisDataSchema,
-      excludedKeys: ['ccbc_number'],
-      overrideParent: 'gis',
-    },
-    application_gis_assessment_hh: {
-      schema: gisAssessmentHhSchema,
-      excludedKeys: [
-        'id',
-        'updated_at',
-        'created_at',
-        'created_by',
-        'updated_by',
-        'archived_at',
-        'archived_by',
-      ],
-      overrideParent: 'gis',
-    },
-    project_information_data: {
-      schema: projectInformationSchema,
-      excludedKeys: [
-        'statementOfWorkUpload',
-        'sowWirelessUpload',
-        'fundingAgreementUpload',
-        'finalizedMapUpload',
-        'otherFiles',
-      ],
-      overrideParent: 'projectInformation',
-    },
-    application_sow_data: {
-      schema: applicationSowDataSchema,
-      excludedKeys: [
-        'province',
-        'ccbc_number',
-        'effectiveStartDate',
-        'projectStartDate',
-        'projectCompletionDate',
-        'backboneFibre',
-        'backboneMicrowave',
-        'backboneSatellite',
-        'lastMileFibre',
-        'lastMileCable',
-        'lastMileDSL',
-        'lastMileMobileWireless',
-        'lastMileFixedWireless',
-        'lastMileSatellite',
-      ],
-      overrideParent: 'application_sow_data',
-    },
-    application_community_progress_report_data: {
-      schema: communityReportSchema,
-      excludedKeys: ['ccbc_number', 'progressReportFile'],
-      overrideParent: 'communityReport',
-    },
-    application_milestone_data: {
-      schema: communityReportSchema,
-      excludedKeys: [
-        'ccbc_number',
-        'milestoneFile',
-        'evidenceOfCompletionFile',
-      ],
-      overrideParent: 'communityReport',
-    },
-    application_pending_change_request: {
-      schema: pendingChangeRequestSchema,
-      excludedKeys: [
-        'id',
-        'created_at',
-        'updated_at',
-        'created_by',
-        'updated_by',
-        'archived_at',
-        'archived_by',
-        'is_pending',
-      ],
-      overrideParent: 'pendingChangeRequest',
-    },
-    application_fnha_contribution: {
-      schema: fnhaContribution,
-      excludedKeys: [
-        'id',
-        'updated_at',
-        'created_at',
-        'created_by',
-        'updated_by',
-        'archived_at',
-        'archived_by',
-        'reason_for_change',
-      ],
-      overrideParent: 'fnhaContribution',
-    },
-    application_communities: {
-      schema: communities.applicationCommunities,
-      excludedKeys: [],
-      overrideParent: 'applicationCommunities',
-    },
-  };
-
-  return configs[tableName] || null;
-};
 
 // Function to map CBC record keys to their sections based on createCbcSchemaData structure
 const getCbcSectionFromKey = (key: string): string => {
